+++ src/Utils/SrGoogleAccountAuthTrait.php
@@ -4,7 +4,6 @@ namespace srag\Plugins\SrGoogleAccountAuth\Utils;
 
 use srag\Plugins\SrGoogleAccountAuth\Access\Access;
 use srag\Plugins\SrGoogleAccountAuth\Access\Ilias;
-use srag\Plugins\SrGoogleAccountAuth\Authentication\Authentication;
 use srag\Plugins\SrGoogleAccountAuth\Client\Client;
 
 /**
@@ -21,14 +20,6 @@ trait SrGoogleAccountAuthTrait {
 	 */
 	protected static function access(): Access {
 		return Access::getInstance();
-	}
-
-
-	/**
-	 * @return Authentication
-	 */
-	protected static function authentication(): Authentication {
-		return Authentication::getInstance();
 	}