+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## [1.1.0]
+- Improvment ILIAS authentication
+- Set external account id
+
 ## [1.0.0]
 - First version